DKEventLoop has been replaced with DKDispatchQueue. Xcode project files have been updated.

+++ DK/DKFramework/Private/Cocoa/View.mm
@@ -317,7 +317,7 @@ using MouseEvent = DKWindow::MouseEvent;
 - (void)postWindowEventType:(WindowEvent::Type)type
 {
 	if (userInstance)
-		userInstance->PostWindowEvent({type, self.windowRect, self.contentRect, self.contentScaleFactor});
+		userInstance->PostWindowEvent({type, self.windowRect, self.contentRect, (float)self.contentScaleFactor});
 }
 
 - (void)postKeyboardEventType:(KeyboardEvent::Type)type keyCode:(unsigned short)keyCode
